+++ CHANGELOG.md
@@ -39,8 +39,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Combined and renamed the following functions:
   - `QDate::from_string` and `QDate::from_string_enum` → `QDate::from_qstring_opt`
   - `QDate::format` and `QDate::format_enum` → `QDate::to_qstring`
-  - `QDateTime::from_date_and_time_time_zone` → `QDateTime::from_date_time_timezone`
-  - `QDateTime::from_date_and_time_time_spec` → `QDateTime::from_date_time_timespec`
+  - `QDateTime::from_date_and_time_time_zone` → `QDateTime::from_qdate_qtime_qtimezone`
+  - `QDateTime::from_date_and_time_time_spec` → `QDateTime::from_qdate_qtime_timespec`
   - `QDateTime::from_string` and `QDateTime::from_string_enum` → `QDateTime::from_qstring_opt`
   - `QDateTime::format` and `QDateTime::format_enum` → `QDateTime::to_qstring`
   - `QImage::from_data` → `QImage::from_data_opt`